% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_if}
\alias{apply_if}
\title{Apply function if condition holds}
\usage{
apply_if(x, cond, f)
}
\arguments{
\item{x}{Object to which function is applied conditionally.}

\item{cond}{Logical.}

\item{f}{Function that takes one argument `x`.}
}
\value{
If `cond` `f(x)`, else `x`.
}
\description{
Apply function if condition holds
}
