% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_latest_release}
\alias{arrow_latest_release}
\title{Filter Dataset for Latest Release}
\usage{
arrow_latest_release(ds, release_column = "release")
}
\arguments{
\item{ds}{A \code{Dataset} as returned by \link{arrow:open_dataset}.}

\item{release_column}{The name of a column which can be interpreted as
a release date. Optimally, the dataset has been partitioned along this
column so that only the latest release data has to be collected in a
subsequent step.}
}
\value{
A \code{Dataset} with filter for \verb{<release_column> == <latest_release>}.
}
\description{
This function determines the latest release of a dataset that has been
partitioned by date and applies a filter for the latest release.
}
