% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_map_chunked}
\alias{arrow_map_chunked}
\title{Apply a function to partitioned data sets}
\usage{
arrow_map_chunked(
  dataset_path_list,
  callback,
  chunk_column_name = "chunk",
  max_index = Inf,
  ...
)
}
\arguments{
\item{dataset_path_list}{A named vector with paths to partitioned
\link[arrow:Dataset]{arrow::Dataset}.}

\item{callback}{A function with at least the arguments \code{ds} and \code{chunk}. \code{ds}
is a named list of data frames. The names are the names of
\code{dataset_path_list}. \code{chunk} is a 1-based sequence counting up.}

\item{chunk_column_name}{A character vector of length 1 or a named character
vector of \code{length(dataset_path_list)} with the same names as
\code{dataset_path_list}. Each element is the column name of the partitioned
column.}

\item{max_index}{Maximum number of combinations of partitioned elements to
iterate over.}

\item{...}{Arguments which are passed to \code{callback}.}
}
\description{
Apply a function to partitioned data sets
}
