% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_vector}
\alias{batch_vector}
\title{Split a vector into sequential batches}
\usage{
batch_vector(x, batch_size)
}
\arguments{
\item{x}{A vector.}

\item{batch_size}{An integer.}
}
\value{
A list where each element contains `batch_size` elements of `x`. The
last element of the list contains at most `batch_size` elements of `x`.
}
\description{
Split a vector into sequential batches
}
