% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster strings}
\usage{
cluster(s, h = 250, dist_method = "lcs", cluster_method = "ward.D")
}
\arguments{
\item{s}{Character vector of strings to be clustered.}

\item{h}{Height at which to cut the tree.}
}
\value{
An integer vector the same length as \code{s} specifying the cluster
assigned to each string.
}
\description{
Cluster strings
}
