% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{cluster_desc}
\alias{cluster_desc}
\title{Cluster description}
\usage{
cluster_desc(
  renv_directory,
  rmd_file,
  iterable,
  iter_to_params,
  output_file = "output.log",
  image_id = "ami-0438747454de030f3",
  instance_type = "t2.small",
  terminate = TRUE,
  security_group_ids = c("sg-01f269087c271cf61", "sg-0f0e8b61aa72dbdef"),
  renv_directory_cmds = NULL,
  renv_directory_cmds_post = NULL
)
}
\description{
Cluster description
}
