% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{compustat_get_financials}
\alias{compustat_get_financials}
\title{Download Compustat Financials}
\usage{
compustat_get_financials(wrds, columns)
}
\arguments{
\item{wrds}{A database connection to WRDS. See \link{wrds_connect}.}

\item{columns}{A character vector containing names of columns to download
from \code{comp_global_daily_all.funda} and \code{comp_global_daily.g_funda}.}
}
\description{
This function downloads financials from the North America and Global
Compustat file.
}
