% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cusip8_to_cusip9}
\alias{cusip8_to_cusip9}
\title{Map 8-digit CUSIP to 9-digit CUSIP}
\usage{
cusip8_to_cusip9(cusip8)
}
\arguments{
\item{cusip8}{A character vector of 8-digit CUSIPs.}
}
\description{
This function computes the check digit and adds it to the 8-digit CUSIP.
}
