% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_search.R
\name{dfs_search}
\alias{dfs_search}
\title{Small DFS Search}
\usage{
dfs_search(
  keywords,
  postback_url = Sys.getenv("POSTBACK_URL"),
  language_code = "en",
  location_code = 2840,
  username = Sys.getenv("DFS_USERNAME"),
  password = Sys.getenv("DFS_PASSWORD")
)
}
\arguments{
\item{keywords}{Character vector of keywords.}

\item{postback_url}{Postback URL to which DataForSEO sends the search results.}

\item{language_code}{Language code.}

\item{location_code}{Location code, see https://api.dataforseo.com/v3/serp/google/locations.}
}
\value{
Output of \link[httr2:req_perform]{httr2::req_perform}.
}
\description{
Search for up to 100 keywords in DataForSEO's Google Organic Search.
}
