% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_search.R
\name{dfs_search_reqs}
\alias{dfs_search_reqs}
\title{Large DFS Search}
\usage{
dfs_search_reqs(
  x,
  task_size = 100,
  postback_url = Sys.getenv("POSTBACK_URL"),
  language_code = "en",
  location_code = 2840,
  username = Sys.getenv("DFS_USERNAME"),
  password = Sys.getenv("DFS_PASSWORD")
)
}
\arguments{
\item{task_size}{Number of keywords that should be put into a single
request. Maximum 100.}

\item{postback_url}{Postback URL to which DataForSEO sends the search results.}

\item{language_code}{Language code.}

\item{location_code}{Location code, see https://api.dataforseo.com/v3/serp/google/locations.}
}
\description{
Search for more than 100 keywords in DataForSEO's Google Organic Search. This
function constructs a list of requests. When submitting these requests, the
calling function has to ensure that the rate limit from less than 2,000
requests with 100 keywords each is met.
}
