% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_patents_view.R
\name{download_patents_view_to_s3}
\alias{download_patents_view_to_s3}
\title{Write a table from PatentsView to S3}
\usage{
download_patents_view_to_s3(table, s3_out)
}
\arguments{
\item{table}{Table name, e.g., `g_location_disambiguated`.}

\item{s3_out}{S3 path, e.g., `s3://<bucket>/path`}
}
\description{
Downloads zipped table from patentsview.org, unzips it, reads it into R
using [readr::read_tsv], and writes it to `s3_out`.
}
