% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_patents_view.R
\name{download_patents_view_to_s3_chunked}
\alias{download_patents_view_to_s3_chunked}
\title{Write a table from PatentsView to S3}
\usage{
download_patents_view_to_s3_chunked(
  table,
  s3_out,
  chunk_size = 1e+05,
  chunk_column_name = "chunk"
)
}
\arguments{
\item{table}{Table name, e.g., \code{g_location_disambiguated}.}

\item{s3_out}{S3 path, e.g., \verb{s3://<bucket>/path}}

\item{chunk_size}{The number of rows to include in each chunk.}

\item{chunk_column_name}{The column name that will be used to partition the
data on S3.}
}
\description{
Downloads zipped table from patentsview.org, unzips it, reads it into R
using \link[readr:read_delim_chunked]{readr::read_tsv_chunked}, and writes each chunk to \code{s3_out}.
}
