% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{ec2_on_demand_limit}
\alias{ec2_on_demand_limit}
\title{EC2 On-Demand Limit}
\usage{
ec2_on_demand_limit()
}
\description{
Returns the service quota for "Running On-Demand Standard (A, C, D, H, I, M,
R, T, Z) instances".
}
