% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ecdf_tbl}
\alias{ecdf_tbl}
\title{Empirical CDF}
\usage{
ecdf_tbl(x, n = 100, fmin = min, fmax = max)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Number of quantiles for which to compute \code{F_q} between \code{fmin(x)} and
\code{fmax(x)}.}

\item{fmin}{Function that returns the minimum quantile.}

\item{fmax}{Function that returns the maximum quantile.}
}
\description{
Create a tibble with two columns: \code{q}: Quantile, \code{F_q}: Empirical CDF at \code{q}.
}
