% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{execute_on_cluster}
\alias{execute_on_cluster}
\title{Execute a parameterized RMD on cluster}
\usage{
execute_on_cluster(
  renv_directory,
  rmd_file,
  iterable,
  iter_to_params,
  image_id = "ami-0438747454de030f3",
  instance_type = "t2.small",
  terminate = TRUE,
  upload_s3_prefix = "upload-rstudio-server-ubuntu-task",
  upload_s3_bucket = "gcpd"
)
}
\arguments{
\item{renv_directory}{Path to a directory on the host machine that is
managed by renv and where all dependencies have been installed into the
AMI image.}

\item{rmd_file}{Path to a RMD script relative to `renv_directory`.}

\item{iterable}{Vector or list. A separate instance is launched for each
element.}

\item{iter_to_params}{Function that transforms an element of `iterable`
into a list of parameters that are passed to `rmarkdown::render` in the
`params` argument.}

\item{image_id}{AMI image id. Each instance will be based on this image.}

\item{instance_type}{EC2 instance type.}

\item{terminate}{If `TRUE` terminate the instance after script completes.}

\item{upload_s3_prefix}{Path to which the JSON file describing the cluster
is uploaded. The default will upload it to a preconfigured folder in S3
that will trigger a lambda which will eventually create the requested
instances.}

\item{upload_s3_bucket}{S3 bucket.}
}
\description{
Execute a parameterized RMD on cluster
}
