% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_patents_view.R
\name{fread_chunked}
\alias{fread_chunked}
\title{Chunked Fread}
\usage{
fread_chunked(file, chunk_size, callback, header = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{file}{File name in working directory, path to file (passed through
\link[base:path.expand]{base::path.expand} for convenience), or a URL starting
http://, file://, etc. Compressed files with extension ‘.gz’ and ‘.bz2’ are
supported if the R.utils package is installed.}

\item{chunk_size}{Number of rows that is read per chunk.}

\item{callback}{A function taking two arguments \code{x} (the data of the current
chunk) and \code{chunk} (the zero-based number of the current chunk). Use this
function to process the chunked data, e.g., by transforming it and writing
it to a database.}

\item{header}{Argument passed to \link[data.table:fread]{data.table::fread}. Currently, only
\code{FALSE} is supported, so that it is suggested to supply \code{col.names} as well.}

\item{verbose}{If \code{TRUE}, print progress information.}

\item{...}{Arguments passed to \link[data.table:fread]{data.table::fread}.}
}
\description{
Chunked version of \link[data.table:fread]{data.table::fread}.
}
