% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_inventory_utils.R
\name{get_most_recent_inventory_date}
\alias{get_most_recent_inventory_date}
\title{Get Most Recent Inventory Date}
\usage{
get_most_recent_inventory_date(bucket = "gcpd")
}
\arguments{
\item{bucket}{S3 bucket.}
}
\description{
Get Most Recent Inventory Date
}
