% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{or_na}
\alias{or_na}
\alias{max_or_na}
\alias{min_or_na}
\alias{mean_or_na}
\alias{sum_or_na}
\title{Return NA if all NA}
\usage{
max_or_na(x)

min_or_na(x)

mean_or_na(x)

sum_or_na(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Variants of standard functions that return \link{base:NA} if all elements of
\code{x} are \code{NA}.
}
