% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{pad_gvkey}
\alias{pad_gvkey}
\title{Pad GVKEY}
\usage{
pad_gvkey(gvkey)
}
\arguments{
\item{gvkey}{A character vector of GVKEYs.}
}
\value{
A character vector of GVKEYs with exactly six digits.
}
\description{
Ensure that GVKEY has six digits. Fill missing digits with 0 from the left.
}
