% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{parallel}
\alias{parallel}
\alias{psum}
\alias{pprod}
\alias{pall}
\alias{pany}
\title{Parallel Computations}
\usage{
psum(..., na.rm = TRUE)

pprod(..., na.rm = TRUE)

pall(..., na.rm = TRUE)

pany(..., na.rm = TRUE)
}
\arguments{
\item{...}{Vectors of the same length.}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\description{
Variants of standard functions that are what \link{base:pmax} is to
\link{base:max}.
}
