% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pasteNA}
\alias{pasteNA}
\alias{paste_NA}
\alias{paste0_NA}
\title{Concatenate Strings}
\usage{
paste_NA(
  ...,
  sep = " ",
  collapse = NULL,
  recycle0 = FALSE,
  NA_if_all_NA = TRUE
)

paste0_NA(..., collapse = NULL, recycle0 = FALSE, NA_if_all_NA = TRUE)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}

\item{recycle0}{\code{\link[base]{logical}} indicating if zero-length
    character arguments should result in the zero-length
    \code{\link[base]{character}(0)}.  Note that when \code{collapse} is
    a string, \code{recycle0} does \emph{not} recycle to zero-length, but
    to \code{""}.}

\item{NA_if_all_NA}{If \code{TRUE}, return \code{NA} if \code{...} only contains \code{NA}. If
\code{FALSE}, return \code{character(0)}.}
}
\description{
Like \link{paste} but \code{NA}s are dropped.
}
