% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patents_view.R
\name{patents_view_assignee_type_map}
\alias{patents_view_assignee_type_map}
\title{PatentsView Assignee Types}
\usage{
patents_view_assignee_type_map()
}
\value{
A named character vector. Names are numerical values (1-9) of
and values are textual descriptions of assignee types.
}
\description{
PatentsView Assignee Types
}
