% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patents_view.R
\name{patents_view_assignee_type_tbl}
\alias{patents_view_assignee_type_tbl}
\title{PatentsView Assignee Types}
\usage{
patents_view_assignee_type_tbl()
}
\value{
A tibble with the following columns:
\itemize{
\item \code{assignee_type} (numeric). PatentsView assignee type.
\item \code{assignee_type_str} (character). Textual description of assignee type.
\item \code{partial_interest} (logical). \code{TRUE} if partial interest.
}
}
\description{
PatentsView Assignee Types
}
