% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patents_view.R
\name{patents_view_last_updated}
\alias{patents_view_last_updated}
\title{PatentsView: Last Updated}
\usage{
patents_view_last_updated()
}
\description{
This function returns the "Last Updated" date for every PatentsView table.
}
