% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patents_view.R
\name{patents_view_urls}
\alias{patents_view_urls}
\title{PatentsView URLs}
\usage{
patents_view_urls()
}
\description{
This function returns the URLs associated with different sets of PatentsView
tables.
}
