% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_nrow}
\alias{print_nrow}
\title{Print number of rows}
\usage{
print_nrow(df, f = identity)
}
\arguments{
\item{df}{Data frame.}

\item{f}{Function. Number of rows of \code{f(df)} is printed.}
}
\value{
The unaltered data frame \code{df}.
}
\description{
Print number of rows
}
