% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_inventory_utils.R
\name{read_inventory}
\alias{read_inventory}
\title{Read Inventory}
\usage{
read_inventory(date = get_most_recent_inventory_date(), bucket)
}
\arguments{
\item{date}{Date in format YYYY-MM-DD.}

\item{bucket}{S3 bucket.}
}
\description{
Read Inventory
}
