% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{s3_key_exists_list}
\alias{s3_key_exists_list}
\title{Apply `s3_key_exists` recursively to S3 keys in a list}
\usage{
s3_key_exists_list(
  l,
  predicate = function(l, name) is.character(l),
  transform = identity,
  verbose = FALSE
)
}
\arguments{
\item{l}{List.}

\item{verbose}{If `TRUE`, print more output.}

\item{is_s3_key}{Function that determines which elements should be treated as S3
keys}
}
\description{
This function raises an error as soon as any S3 key does not exist.
}
