% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{s3_list_subfolders}
\alias{s3_list_subfolders}
\title{List S3 Subfolders}
\usage{
s3_list_subfolders(
  prefix = NULL,
  bucket = "gcpd",
  max = 1000,
  subfolder_only = FALSE
)
}
\arguments{
\item{prefix}{Common prefix to all files.}

\item{bucket}{S3 bucket.}

\item{max}{Maximum number of results to return.}

\item{subfolder_only}{Remove the prefix from the subfolder S3 key.}
}
\description{
Subfolders are exactly one level further down than the given prefix.
}
