% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{s3_put_object}
\alias{s3_put_object}
\title{Upload file to S3}
\usage{
s3_put_object(file, s3_key, bucket = "gcpd")
}
\arguments{
\item{file}{File.}

\item{s3_key}{S3 key.}

\item{bucket}{S3 bucket.}
}
\description{
Upload file to S3
}
