% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{s3_read_json}
\alias{s3_read_json}
\title{Read JSON file from S3}
\usage{
s3_read_json()
}
\arguments{
\item{bucket}{S3 bucket.}

\item{key}{S3 key of JSON file.}
}
\value{
A list representing the contents of the JSON file.
}
\description{
Read JSON file from S3
}
