% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{save_figure}
\alias{save_figure}
\title{Save Figure}
\usage{
save_figure(p, file, file_formats, width = 7, height = 5, ...)
}
\arguments{
\item{p}{Plot to save.}

\item{file}{File name to create on disk. Make sure to not include the file
type, i.e., if you want to create a PDF do not put ".pdf" at the end.}

\item{file_formats}{A character vector of file formats supported by
\link[ggplot2:ggsave]{ggplot2::ggsave}. The figure is saved for each of these file formats.}

\item{width, height}{Plot size in units expressed by the \emph{units} argument. If
not supplied, uses the size of the current graphics device.}
}
\description{
Interface to \link[ggplot2:ggsave]{ggplot2::ggsave}.
}
