% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_non_na}
\alias{unique_non_na}
\title{Extract Unique Elements (Without NA)}
\usage{
unique_non_na(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{a vector or a data frame or an array or \code{NULL}.}

\item{incomparables}{a vector of values that cannot be compared.
    \code{FALSE} is a special value, meaning that all values can be
    compared, and may be the only value accepted for methods other than
    the default.  It will be coerced internally to the same type as
       \code{x}.}

\item{...}{arguments for particular methods.}
}
\description{
Like \link{unique} but \code{NA}s are dropped.
}
