% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize}
\usage{
winsorize(x, probs)
}
\arguments{
\item{x}{Numerical vector.}

\item{probs}{A numerical vector of length 2 specifying the quantiles at which
to winsorize. Use \code{0} or \code{1} if you do not want to winsorize at the lower end
or the upper end, respectively.}
}
\description{
Winsorize a numerical vector.
}
