% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{winsorize_bounds}
\alias{winsorize_bounds}
\title{Winsorize Bounds}
\usage{
winsorize_bounds(x, probs)
}
\arguments{
\item{x}{Numerical vector.}

\item{probs}{A numerical vector of length 2 specifying the quantiles at which
to winsorize. Use \code{0} or \code{1} if you do not want to winsorize at the lower end
or the upper end, respectively.}
}
\value{
A numerical vector of length 2. The first element is the lower bound
and the second element is the upper bound.
}
\description{
Get winsorize bounds for a numerical vector. This function is effectively
a wrapper around \link{stats:quantile}.
}
