% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_connect}
\alias{wrds_connect}
\title{Connect to WRDS}
\usage{
wrds_connect(user)
}
\arguments{
\item{user}{Character. WRDS user name.}
}
\value{
A `DBI` connection to WRDS.
}
\description{
Make sure to setup your [local RStudio](https://wrds-www.wharton.upenn.edu/pages/support/programming-wrds/programming-r/r-from-the-web/)
or [RStudioServer](https://wrds-www.wharton.upenn.edu/pages/support/programming-wrds/programming-r/r-from-the-web/)
as described on the linked websites.
}
