% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_dataset_count}
\alias{wrds_dataset_count}
\title{WRDS Dataset Count}
\usage{
wrds_dataset_count(conn, library, dataset)
}
\arguments{
\item{conn}{Connection to WRDS from [wrds_connect()].}

\item{library}{Name of WRDS library, e.g., `"crsp"`.}

\item{dataset}{Name of WRDS dataset within the library.}
}
\value{
An integer.
}
\description{
Get number of records in a WRDS dataset.
}
