% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_datasets}
\alias{wrds_datasets}
\title{WRDS Datasets}
\usage{
wrds_datasets(conn, library)
}
\arguments{
\item{conn}{Connection to WRDS from [wrds_connect()].}

\item{library}{Name of WRDS library, e.g., `"crsp"`.}
}
\value{
A tibble with column `table_name`.
}
\description{
Get all available datasets within a given WRDS library.
}
