% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_libraries}
\alias{wrds_libraries}
\title{WRDS Libraries}
\usage{
wrds_libraries(conn)
}
\arguments{
\item{conn}{Connection to WRDS from [wrds_connect()].}
}
\value{
A tibble with column `table_schema`.
}
\description{
Get a list of all available WRDS libraries.
}
