% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_schema_tables}
\alias{wrds_schema_tables}
\title{WRDS Tables}
\usage{
wrds_schema_tables(conn, schema)
}
\arguments{
\item{conn}{Database connection to WRDS created with \link{wrds_connect}.}

\item{schema}{Name of WRDS schema, e.g., \code{"crsp"}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{table_schema}: \code{schema}.
\item \code{table_name}: Table name. Use \link{wrds_table_columns} to query column names
of a table.
\item \code{table_type}:  Table type. One of \code{"BASE TABLE"}, \code{"VIEW"} or
\code{"FOREIGN TABLE"}.
}
}
\description{
Get all available tables under a schema.
}
