% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_schemata}
\alias{wrds_schemata}
\title{WRDS Schemata}
\usage{
wrds_schemata(conn)
}
\arguments{
\item{conn}{Database connection to WRDS created with \link{wrds_connect}.}
}
\value{
A tibble with the following column:
\itemize{
\item \code{schema_name}: Schema name. Use \link{wrds_schema_tables} to query all tables
under a schema.
}
}
\description{
Get all available WRDS schemata.
}
