% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_table_columns}
\alias{wrds_table_columns}
\title{WRDS Table Columns}
\usage{
wrds_table_columns(conn, schema, table)
}
\arguments{
\item{conn}{Database connection to WRDS created with \link{wrds_connect}.}

\item{schema}{Name of WRDS schema, e.g., \code{"crsp"}.}

\item{table}{Table name within the given schema.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{table_schema}: \code{schema}.
\item \code{table_name}: \code{table}.
\item \code{column_name}: Column name.
\item \code{column_default}: Default value.
\item \code{is_nullable}: \code{"YES"} or \code{"NO"}.
\item \code{data_type}: \href{https://www.postgresql.org/docs/current/datatype.html}{Data Type}.
}
}
\description{
Get all columns of a WRDS table.
}
