% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrds_table_count}
\alias{wrds_table_count}
\title{WRDS Table Count}
\usage{
wrds_table_count(conn, schema, table)
}
\arguments{
\item{conn}{Database connection to WRDS created with \link{wrds_connect}.}

\item{schema}{Name of WRDS schema, e.g., \code{"crsp"}.}

\item{table}{Table name within the given schema.}
}
\value{
An integer.
}
\description{
Get number of records in a WRDS table
}
