% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{write_json_to_s3}
\alias{write_json_to_s3}
\title{Write JSON to S3}
\usage{
write_json_to_s3(x, s3_key, bucket = "gcpd")
}
\arguments{
\item{x}{List.}

\item{s3_key}{S3 key.}

\item{bucket}{S3 bucket.}
}
\description{
Write JSON to S3
}
