% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{cluster_desc}
\alias{cluster_desc}
\title{Cluster description}
\usage{
cluster_desc(
  renv_directory,
  rmd_file,
  iterable,
  iter_to_params,
  image_id = "ami-0438747454de030f3",
  instance_type = "t2.small"
)
}
\description{
Cluster description
}
