% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{compustat_annualize_financials}
\alias{compustat_annualize_financials}
\title{Annualize Compustat Financials}
\usage{
compustat_annualize_financials(financials, flow_variables, stock_variables)
}
\arguments{
\item{financials}{Compustat financials as returned by
\link{compustat_get_financials}.}

\item{flow_variables}{Character vector of flow variables. For example,
\code{"capx"} or \code{"revt"}.}

\item{stock_variables}{Character vector of stock variables. For example,
\code{"at"} or \code{"csoh"}.}
}
\description{
Annualize Compustat Financials
}
\details{
Assumptions:
\itemize{
\item Flow variables are spread uniformly across months, i.e., if a firm files
their annual report at the end of march, 3/12 of a flow variable is
counted for t and 9/12 is counted for t-1.
\item Flow variables are reported for the period between previous fiscal year end
and current fiscal year end, i.e., if a company changes the month of their
fiscal year from June in 1994 to March in 1995 (as was the case for Toyota), the
flow only accrues for the nine months between July 1994 and March 1995.
TODO: THERE MIGHT BE SOME FIRMS THAT REPORT THE PREVIOUS 12 MONTH FLOW. COULD
BE DETECTED BY FILTERING FOR FISCAL YEAR END CHANGES AND LOOKING FOR UNUSUALLY
BIG JUMPS IN FLOW VARIABLES (COULD LOOK AT CORRELATION ACROSS ALL FLOW VARIABLES
FOR HIGHER CERTAINTY)
\item Stock variables grow linearly between fiscal year end of t-1 and t. Then,
stock at end of calendar year can be determined by linear interpolation.
\item Each month length is 1/12 of year length
}

Need to check what happens if month of fiscal year end changes for a given firm
}
