% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patents_view_schema.R
\name{compute_patents_view_schema}
\alias{compute_patents_view_schema}
\title{Get PatentsView Schema}
\usage{
compute_patents_view_schema(
  data_download_dictionary_url =
    "https://patentsview.org/download/data-download-dictionary"
)
}
\description{
This function converts the data download dictionary from PatentsView on
https://patentsview.org/download/data-download-dictionary into a tibble.
}
