% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_patents_view.R
\name{download_patents_view}
\alias{download_patents_view}
\title{Read a table from PatentsView}
\usage{
download_patents_view(table)
}
\arguments{
\item{table}{Table name, e.g., `g_location_disambiguated`.}
}
\value{
Path to the unzipped tsv file.
}
\description{
Downloads zipped table from patentsview.org, unzips it, and returns the path
to the unzipped file.
}
