% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_split}
\alias{index_split}
\title{Split a vector into modulus batches}
\usage{
index_split(x, n)
}
\arguments{
\item{x}{A vector.}

\item{n}{An integer.}
}
\value{
A list with at most \code{n} elements. The \code{i}-th element is a vector
containing all the elements of \code{x} belonging to the same equivalence class
with respect to the \code{modulus} operation. That means \eqn{x_1} and \eqn{x_2}
are both elements of \code{out[i]} if eqn{x_1 mod n = x_2 mod n = i}.
}
\description{
Split a vector into modulus batches
}
