% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_rmd_params.R
\name{parse_year_params}
\alias{parse_year_params}
\title{Parse RMarkdown Year Parameters}
\usage{
parse_year_params(params)
}
\arguments{
\item{params}{The \code{params} object made available by RMarkdown. It should
have the fields \code{years} (a comma-separated string of years), \code{year_start}
or \code{year_end}. If \code{years} is different from \code{NULL} (\code{~} in YAML), \code{years}
takes precedence over \code{year_start} and \code{year_end}.}
}
\value{
An integer vector of years.
}
\description{
This is a convenience function to determine the set of years for which the
RMarkdown script has to be run. This is useful when the RMarkdown script
independently executes a function for each year.
}
