% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_if}
\alias{print_if}
\title{Print object if condition holds}
\usage{
print_if(x, cond)
}
\arguments{
\item{x}{Object to be printed conditionally.}

\item{cond}{Logical. If \code{TRUE}, \code{x} is printed.}
}
\value{
The unaltered object \code{x}.
}
\description{
Print object if condition holds
}
