% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{s3_key_exists_list}
\alias{s3_key_exists_list}
\title{Apply \code{s3_key_exists} recursively to S3 keys in a list}
\usage{
s3_key_exists_list(
  l,
  predicate = function(l, name) is.character(l),
  transform = identity,
  verbose = FALSE
)
}
\arguments{
\item{l}{List.}

\item{predicate}{A function taking two arguments \code{l} (an element of a list)
and \code{name} (currently defunct). If this function returns \code{TRUE}, \code{transform}
is applied on this element of the list.}

\item{transform}{A function taking one argument \code{l} (an element of a list).
This function should return a character representing a potential S3 key. This
character is then passed to \link{s3_key_exists}.}

\item{verbose}{If \code{TRUE}, print more output.}
}
\description{
This function raises an error as soon as any S3 key does not exist.
}
