% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.R
\name{s3_list_files}
\alias{s3_list_files}
\title{List Files on S3}
\usage{
s3_list_files(
  prefix = NULL,
  bucket = "gcpd",
  include_folder = FALSE,
  max = 1000,
  delimiter = NULL
)
}
\arguments{
\item{prefix}{Common prefix to all files.}

\item{bucket}{S3 bucket.}

\item{include_folder}{If \code{TRUE}, also include S3 keys referring to folders.}

\item{max}{Maximum number of results to return.}

\item{delimiter}{Character string used to group keys. Read the AWS docs for
more details.}

\item{value}{See \code{Value}.}
}
\value{
If \code{value == "vector"}, a character vector of S3 keys. If
\code{value == "df"} a data frame of S3 keys.
}
\description{
List Files on S3
}
